added @TODO: fix results
there's a bug with multiple indices (which is the core product here).
I'm iterating too many times over different divs because I thought
setting up $results to be set on every search was a good way to handle
it.

would probably be best to collect that data and save it in the helpers[]
array. This way we can always have reference to the $result element
without having to rebuild the variable each pass.