big fixes to $results bug
changed the methodology of how I assign the indices to each results
area.

this time I’m using the .each() functionality of the plugin integration
to tie the plugin functions to each element associated with the query
selector.

by doing this I can store the index, hitsPerPage, and any infinite
other options as data-* attribs on the dom level.

this also mimics angular ng-* dom level methodology so the workflow
becomes easier as well.

Algolia is the Model.
HTML is the view.
FancyAlgolia.js is the controller

self pat on the back